test: dedupe diff coverage and share harness
